+++ NEWS.md
@@ -7,3 +7,12 @@ arcos 0.82
 -------------------------------------------------------
 
 Updated to follow new API path to account for future versions (ie /v1/county_data/)
+
+arcos 1.0
+-------------------------------------------------------
+
+Updated to include newly released 2013 and 2014 data.
+
+Added nearly 10,000 more geocoded pharmacy locations and fixed a few dozen latitudes and longitudes.
+
+New functions to query list of counties, drug types, business types, and raw data for individual drugs by county.